refactor: 컨트롤러를 템플릿 컨트롤러와 Rest 컨트롤러로 분리
