+++ src/main/java/roomescape/controller/RoomEscapeCommandController.java
@@ -2,45 +2,33 @@ package roomescape.controller;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.Reservation;
 import roomescape.domain.dto.ReservationRequestDto;
 import roomescape.service.RoomEscapeService;
-import roomescape.domain.Reservation;
 
-@Controller
-public class RoomEscapeController {
+@RestController
+public class RoomEscapeCommandController {
 
-    private List<Reservation> reservations = new ArrayList<>();
     private final RoomEscapeService roomEscapeService;
+    private List<Reservation> reservations = new ArrayList<>();
 
-    public RoomEscapeController(RoomEscapeService roomEscapeService) {
+    public RoomEscapeCommandController(RoomEscapeService roomEscapeService) {
         this.roomEscapeService = roomEscapeService;
     }
 
-    @GetMapping("admin")
-    public String home() {
-        return "admin/index";
-    }
-
-    @GetMapping("admin/reservation")
-    public String reservation() {
-        return "admin/reservation-legacy";
-    }
-
     @GetMapping("reservations")
-    @ResponseBody
     public List<Reservation> readReservations() {
         return reservations;
     }
 
     @PostMapping("reservations")
-    @ResponseBody
     public Reservation add(@RequestBody ReservationRequestDto reservationDto) {
         Reservation reservation = roomEscapeService.convertReservation(reservationDto);
         reservations.add(reservation);
@@ -48,7 +36,6 @@ public class RoomEscapeController {
     }
 
     @DeleteMapping("reservations/{reservationId}")
-    @ResponseBody
     public void delete(@PathVariable("reservationId") Long id) {
         Reservation reservation = reservations.stream()
                 .filter(v -> v.getId().equals(id))
